+++ src/main/java/gift/domain/model/Product.java
@@ -1,4 +1,4 @@
-package gift;
+package gift.domain.model;
 
 public class Product {
